requirements.txt
asgiref==3.2.3
Django==3.0.3
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
python-magic==0.4.15
python-magic-bin==0.4.14
python-tds==1.10.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
Werkzeug==1.0.0
